Created `gaia_techne.jl` with the provided code and verified its creation, replacing the old update script.
