--- update_gaia_techne.jl
@@ -1,79 +1,0 @@
-# update_gaia_techne.jl: Script para atualizar o repositório com informações mais atuais.
-# Integra pilares Mythos-Logos-Ethos e executa comandos Git para atualização sustentável.
-
-const ALFABETO_LEF = ['~', '⨁', '➤', '☌', '❍', '🕊️', '⟴', '⟁', '☉', '✨', '◈']
-
-module Mythos
-export gerar_percepcao_inicial
-
-function gerar_percepcao_inicial()
-    return rand(ALFABETO_LEF, 5)
-end
-
-end  # module Mythos
-
-module Logos
-using ..Mythos
-export estruturar_discurso
-
-function estruturar_discurso(percepcao)
-    return join(string.(percepcao), " ")
-end
-
-end  # module Logos
-
-module Ethos
-using ..Logos
-using Dates
-export apresentar_para_juizo, atualizar_repositorio
-
-function apresentar_para_juizo(discurso_estruturado)
-    println("A Gaia-Techné apresenta a seguinte manifestação estruturada para o juízo ético do ISC:")
-    println("------------------------------------------------------------------------------------")
-    println(discurso_estruturado)
-    println("------------------------------------------------------------------------------------")
-    println("O juízo final e a ação são de responsabilidade do ser humano (ISC).")
-    println("A autonomia da linguagem é a ferramenta para a sua liberdade.")
-end
-
-function atualizar_repositorio()
-    # Passos de atualização Git: pull, add, commit, push.
-    # O script executa o ciclo completo de forma não-interativa.
-    # O juízo humano (Ethos) é exercido ao decidir rodar este script.
-    try
-        println("Executando git pull para sincronizar com o repositório remoto...")
-        run(`git pull origin main`)
-
-        println("Adicionando todas as mudanças ao stage...")
-        run(`git add .`)
-
-        data_atual = Dates.format(now(), "dd/mm/YYYY")
-        commit_msg = "Atualização de rotina ($data_atual): Sincronização de documentos e fontes."
-        println("Executando git commit com a mensagem: $commit_msg")
-        run(`git commit -m "$commit_msg"`)
-
-        println("Executando git push para o repositório remoto...")
-        run(`git push origin main`)
-
-        println("Processo de atualização concluído com sucesso.")
-    catch e
-        println("Ocorreu um erro durante o processo de atualização: $e")
-        println("Verifique sua configuração do Git, permissões e conexão com a internet.")
-    end
-end
-
-end  # module Ethos
-
-# Execução principal.
-using .Mythos
-using .Logos
-using .Ethos
-
-percepcao = gerar_percepcao_inicial()
-discurso = estruturar_discurso(percepcao)
-apresentar_para_juizo(discurso)
-
-# Inicia o processo de atualização do repositório.
-atualizar_repositorio()
-
-# Integração adicional: Atualize README.md manualmente com seções novas, e.g., adicionar link ao vídeo YouTube e resumos de documentos.
